+++ doc/CHANGES.md
@@ -5,6 +5,7 @@ Please see [rsbackup in git](https://github.com/ewxrjk/rsbackup) for detailed ch
 ## Changes In rsbackup 10.0
 
 * Pass `--open-noatime` to rsync by default. Fixes [issue #101](https://github.com/ewxrjk/rsbackup/issues/101).
+* **Incompatible change**: `rsync-timeout`, the old name for `backup-job-timeout`, has been removed.
 
 ## Changes In rsbackup 9.0